LayoutTests/imported/w3c/web-platform-tests/cookies/prefix/__HostHttp.https-expected.txt


PASS __HostHttp: Does not set via DOM 'Secure; Path=/'
PASS __HostHttp: Does not set via DOM with Domain attribute 'Secure; Path=/; Domain=localhost'
FAIL __HostHttp: Does not set via HTTP with 'Secure; Path=/' assert_equals: expected (undefined) undefined but got (string) "bar1"
PASS __HostHttp: Set via HTTP with 'Secure; Path=/; httponly'
FAIL __HostHttp: Does not set via HTTP with 'Secure; Path=/cookies/; httponly' assert_equals: expected (undefined) undefined but got (string) "bar3"
FAIL __HostHttp: Does not set via HTTP with 'Path=/;' (without Secure) assert_equals: expected (undefined) undefined but got (string) "bar4"
FAIL __HostHttp: Does not set via HTTP with Domain attribute 'Secure; Path=/; Domain=localhost' assert_equals: expected (undefined) undefined but got (string) "bar5"
